Split up styles
