+++ src/TerminalPage/components/Path.tsx
@@ -1,4 +1,4 @@
-import styles from '../TerminalPage.module.scss';
+import styles from '../styles/path.module.scss';
 import {Time} from "./Time.tsx";
 import {useContext} from "react";
 import {AppContext} from "../AppContext.tsx";
