+++ src/TerminalPage/styles/TerminalPage.module.scss
@@ -1,0 +1,21 @@
+html, body {
+  background-color: black;
+  color: #00ff00;
+  font-family: 'VT323', monospace;
+  font-size: 1.5rem;
+  height: 100%;
+  line-height: 1.75rem;
+  margin: 0;
+  padding: 0;
+}
+
+.terminal_container {
+  border: solid 2px #00ff00;
+  display: flex;
+  flex-direction: column;
+  height: 80vh;
+  justify-content: space-between;
+  margin: auto;
+  width: 80%;
+}
+
