+++ src/TerminalPage/styles/display.module.scss
@@ -1,0 +1,9 @@
+.display {
+  display: flex;
+  height: 100%;
+}
+
+.display_pane {
+  flex: 0 0 75%;
+  width: 100%;
+}
