+++ src/TerminalPage/styles/filesystem.module.scss
@@ -1,44 +1,3 @@
-// TODO: Split this file up
-
-html, body {
-  background-color: black;
-  color: #00ff00;
-  font-family: 'VT323', monospace;
-  font-size: 1.5rem;
-  height: 100%;
-  line-height: 1.75rem;
-  margin: 0;
-  padding: 0;
-}
-
-.terminal_container {
-  border: solid 2px #00ff00;
-  display: flex;
-  flex-direction: column;
-  height: 80vh;
-  justify-content: space-between;
-  margin: auto;
-  width: 80%;
-}
-
-.display {
-  display: flex;
-  height: 100%;
-}
-
-.path {
-  border-top: solid #00ff00 2px;
-  display: flex;
-  flex: 0 0 1rem;
-  justify-content: space-between;
-  padding: 0 0.25rem 0 0.25rem;
-}
-
-.display_pane {
-  flex: 0 0 75%;
-  width: 100%;
-}
-
 .filesystem {
   border-right: solid #00ff00 2px;
   display: flex;
@@ -81,13 +40,5 @@ html, body {
 
 .selected {
   background-color: #00ff00;
-  color: black;
-}
-
-.tick {
-  color: #00ff00;
-}
-
-.tock {
   color: black;
 }
