+++ src/TerminalPage/styles/path.module.scss
@@ -1,0 +1,15 @@
+.path {
+  border-top: solid #00ff00 2px;
+  display: flex;
+  flex: 0 0 1rem;
+  justify-content: space-between;
+  padding: 0 0.25rem 0 0.25rem;
+}
+
+.tick {
+  color: #00ff00;
+}
+
+.tock {
+  color: black;
+}
